+++ loans/src/main/java/com/eazybytes/loans/entity/Loans.java
@@ -24,9 +24,9 @@ public class Loans extends BaseEntity {
 
     private String loanType;
 
-    private Long totalLoan;
+    private int totalLoan;
 
-    private Long amountPaid;
+    private int amountPaid;
 
-    private Long outstandingAmount;
+    private int outstandingAmount;
 }
